proactive-tests/exec_get_zenoh_file/requirements.txt
eclipse-zenoh==0.11.0
maturin>=0.13,<0.15